+++ src/org/sam/colisiones/Segmento.java
@@ -7,20 +7,20 @@ package org.sam.colisiones;
  */
 public class Segmento {
 
-	double x1, y1, x2, y2;
+	float x1, y1, x2, y2;
 	
 	public Segmento(){
-		x1 = y1 = x2 = y2 = 0.0;
+		x1 = y1 = x2 = y2 = 0.0f;
 	}
 	
-	public Segmento(double X1, double Y1, double X2, double Y2){
+	public Segmento(float X1, float Y1, float X2, float Y2){
 		x1 = X1;
 		y1 = Y1;
 		x2 = X2;
 		y2 = Y2;
 	}
 	
-	public void setValues(double X1, double Y1, double X2, double Y2){
+	public void setValues(float X1, float Y1, float X2, float Y2){
 		x1 = X1;
 		y1 = Y1;
 		x2 = X2;
@@ -36,7 +36,7 @@ public class Segmento {
 	 * @return La distancia de un punto a otro
 	 */
 
-	public static double distancia(double X1, double Y1, double X2, double Y2) {
+	public static double distancia(float X1, float Y1, float X2, float Y2) {
 		X2 -= X1; // se almacena en X2 incremento de X
 		Y2 -= Y1; // se almacena en Y2 incremento de Y
 		return  Math.sqrt(X2 * X2 + Y2 * Y2);
@@ -76,21 +76,21 @@ public class Segmento {
  	 * 		 2 si el punto esta por encima de la linea del segmento 
 	 */
 
-	private static int posicionRelaivaAlSegmento(double X1, double Y1,
-			double X2, double Y2,
-			double PX, double PY) {
+	private static int posicionRelaivaAlSegmento(float X1, float Y1,
+			float X2, float Y2,
+			float PX, float PY) {
 		
 		X2 -= X1; // se almacena en X2 incremento de X
 		Y2 -= Y1; // se almacena en Y2 incremento de Y
 		PX -= X1; // la posicion del punto en X se hace realitiva al primer punto
 		PY -= Y1; // la posicion del punto en Y se hace realitiva al primer punto
 
-		double prs = Y2 * PX - X2 * PY;
+		float prs = Y2 * PX - X2 * PY;
 		return (prs < 0.0) ? 1 : ((prs > 0.0) ? 2 : 0);
 	}
 
 	/**
-	 * Coprueba si el segmento  (X1,Y1)(X2,Y2) corta al segmento (X1,Y1)(X2,Y2) 
+	 * Coprueba si el segmento  (X1,Y1)(X2,Y2) corta al segmento (X3,Y3)(X4,Y4) 
 	 *
 	 * @param X1,&nbsp;Y1 Coordenadas del primer punto del primer segmento
 	 * @param X2,&nbsp;Y2 Coordenadas del segundo punto del primer segmento
@@ -103,10 +103,10 @@ public class Segmento {
 	 *		false en caso contrario
 	 */
 
-	public static boolean hayInterseccion(double X1, double Y1,
-			double X2, double Y2,
-			double X3, double Y3,
-			double X4, double Y4) {
+	public static boolean hayInterseccion(float X1, float Y1,
+			float X2, float Y2,
+			float X3, float Y3,
+			float X4, float Y4) {
 
 		return  ((	posicionRelaivaAlSegmento(X1, Y1, X2, Y2, X3, Y3) &
 					posicionRelaivaAlSegmento(X1, Y1, X2, Y2, X4, Y4)	) == 0)
@@ -115,7 +115,7 @@ public class Segmento {
 					posicionRelaivaAlSegmento(X3, Y3, X4, Y4, X2, Y2)	) == 0);
 	}
 	
-	public boolean hayInterseccion(double X3, double Y3, double X4, double Y4){
+	public boolean hayInterseccion(float X3, float Y3, float X4, float Y4){
 		return  hayInterseccion(x1, y1, x2, y2, X3, Y3, X4, Y4);
 	}
